requirements.txt
pandas
pandas-stubs
plotly
loguru
jmespath
types-jmespath
loguru
numpy
pytest
geopy
scikit-learn
python-dotenv
scrapfly-sdk
scrapy
sqlalchemy
mysql-connector-python
matplotlib
seaborn
mypy